+++ packages/cedar-amcharts/src/index.ts
@@ -2,7 +2,7 @@ import render from './render/render'
 
 export function cedarAmCharts(elementId: string, definition: any, data?: any[]) {
   if ((!elementId || !definition || !data) && (definition.type && definition.type !== 'custom')) {
-    const err = new Error('Element Id, definitionification, and data are all required.')
+    const err = new Error('Element Id, definition, and data are all required.')
     throw err
   }