+++ packages/cedar-amcharts/CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - legend.enable is now legend.visible
 - Text colors have been changed to pass color contrast ratio on white backgrounds
+### Fixed
+- Specifications that are passed in are actually supported
 
 ## 1.0.0-beta.1
 ### Added